Show filters icon only when there are some variables to show
