+++ public/app/features/dashboard/components/SubMenu/SubMenuItems.tsx
@@ -1,4 +1,6 @@
-import React, { useEffect, useState } from 'react';
+import FilterListIcon from '@mui/icons-material/FilterList';
+import { Box, styled } from '@mui/material';
+import React, { FC, useEffect, useState } from 'react';
 
 import { selectors } from '@grafana/e2e-selectors';
 import { useSelector } from 'app/types';
@@ -17,8 +19,10 @@ export const SubMenuItems = ({ variables, readOnly }: Props) => {
   const hiddenVariables = useSelector((state) => state.fnGlobalState.hiddenVariables);
 
   useEffect(() => {
-    setVisibleVariables(variables.filter((state) => state.hide !== VariableHide.hideVariable));
-  }, [variables]);
+    setVisibleVariables(
+      variables.filter((state) => state.hide !== VariableHide.hideVariable && !hiddenVariables?.includes(state.id))
+    );
+  }, [variables, hiddenVariables]);
 
   if (visibleVariables.length === 0) {
     return null;
@@ -26,11 +30,8 @@ export const SubMenuItems = ({ variables, readOnly }: Props) => {
 
   return (
     <>
+      <FilterWithIcon />
       {visibleVariables.map((variable) => {
-        if (hiddenVariables?.includes(variable.id)) {
-          return null;
-        }
-
         return (
           <div
             key={variable.id}
@@ -44,3 +45,20 @@ export const SubMenuItems = ({ variables, readOnly }: Props) => {
     </>
   );
 };
+
+const FilterWithIcon: FC = () => (
+  <FilterWithIconStyled>
+    <FilterListIcon sx={{ color: '#3A785E' }} />
+    FILTERS
+  </FilterWithIconStyled>
+);
+
+const FilterWithIconStyled = styled(Box)({
+  display: 'flex',
+  gap: 1,
+  alignItems: 'center',
+  color: '#3A785E',
+  fontWeight: 600,
+  lineHeight: '160%',
+  fontSize: 12,
+});
